requirements.txt
requests
beautifulsoup4
pandas
sqlalchemy
pytest
pyyaml